Splitting Calendar class to its own file
